Move plotting to separate file
